requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
extruct==0.18.0
flexcache==0.3
flexparser==0.4
html-text==0.7.1
html5lib==1.1
idna==3.11
ingredient_parser_nlp==2.4.0
isodate==0.7.2
joblib==1.5.3
jstyleson==0.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
mf2py==2.0.1
nltk==3.9.2
numpy==2.4.0
Pint==0.25.2
platformdirs==4.5.1
pyparsing==3.3.1
pyRdfa3==3.6.4
python-crfsuite==0.9.11
rdflib==7.5.0
recipe_scrapers==15.11.0
regex==2025.11.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
w3lib==2.3.1
webencodings==0.5.1
